requirements.txt
faiss-cpu
jq
Flask
telebot
langchain
langchain_openai
langchain_text_splitters
langchain_community.vectorstores
langsmith
langdetect
nltk
openai
packaging
pandas
requests
tenacity
python-dotenv
deep_translator
logging
flask
soundfile

